backend/requirements.txt
# FastAPI & Web Server
fastapi[standard]
uvicorn

# Database
sqlalchemy
sqlmodel
asyncpg
aiosqlite
alembic
psycopg2

# Validation & Settings
pydantic
pydantic-settings
python-dotenv

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]
PyJWT

# Testing
pytest
pytest-asyncio
pytest-cov
httpx

# Utilities
requests
aiofiles
email-validator

# Mercado Pago
mercadopago

# Rate limiting
slowapi

# Optional but recommended
black
flake8
mypy
isort